refactor: combat engine turn loop replaces runSinglePass; dpsSimulator becomes adapter (parity-locked)
